move abi-gen funcs from index to utils for testing
preparing for unit testing. purely refactoring (no functionality
changed).
